requirements/base.txt
django
django-environ
djangorestframework
django-cors-headers
django-filter
django-autoslug
django-countries
django-phonenumber-field
phonenumbers
drf-yasg
Pillow
argon2-cffi
pytz
django-taggit
celery
redis
flower
django-celery-email
django-allauth
djangorestframework-simplejwt
dj-rest-auth
django-filter